created a separate SearchTable component and moved all of the logic that was previously within App.tsx, now instead of rendering SearchBar and Table components separately inside of App.tsx, I'm only rendering the SearchTable component
