+++ src/components/SearchTable/index.ts
@@ -1,0 +1,2 @@
+import SearchTable from './SearchTable';
+export default SearchTable;
