+++ src/components/SearchTable/style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const StyledSearchTable = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 52px;
+`;
